Extract the number of suppressed warnings
It is hard for a user to see the total number of suppressed warnings in
their source code. It should be easier with this functionality which
extracts it from each clang-tidy run and sums it up.

To discourage users from suppressing warnings all over their code base
they are slightly penalized for each found suppression.